add tests for help and split SVM tests
