+++ src/SimpleAccounting/Infrastructure/Defines.cs
@@ -4,6 +4,9 @@
 
 namespace lg2de.SimpleAccounting.Infrastructure
 {
+    using System.Diagnostics.CodeAnalysis;
+
+    [ExcludeFromCodeCoverage]
     internal static class Defines
     {
         public const string ProjectName = "SimpleAccounting";
@@ -12,6 +15,5 @@ namespace lg2de.SimpleAccounting.Infrastructure
 
         public static readonly string ProjectUrl = $"https://{GithubDomain}/{OrganizationName}/{ProjectName}";
         public static readonly string NewIssueUrl = $"{ProjectUrl}/issues/new?template=bug-report.md";
-
     }
 }
